vista lista trabajador

+++ src/main/java/com/example/lab2/Controller/TrabajadorController.java
@@ -1,5 +1,6 @@
 package com.example.lab2.Controller;
 
+import com.example.lab2.Entity.Tipo;
 import com.example.lab2.Entity.Trabajador;
 import com.example.lab2.Repository.TrabajadorRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,45 +15,52 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import java.util.List;
 import java.util.Optional;
 
+
 @Controller
 @RequestMapping(value = "trabajador")
 public class TrabajadorController {
     @Autowired
     TrabajadorRepository trabajadorRepository;
 
-    @GetMapping(value = "trabajador/lista")
+    @GetMapping(value = "lista")
     public String listar(Model model){
-        List<Trabajador> trabajador = trabajadorRepository.findAll();
-        model.addAttribute("lista",trabajador);
+        List<Trabajador> trabajadorLista = trabajadorRepository.findAll();
+        model.addAttribute("lista",trabajadorLista);
+        System.out.println(trabajadorLista);
         return "trabajador/lista";
 
     }
 
-    @GetMapping({"/trabajador/nuevo"})
-    public String nuevoTrabajador() {
+    @GetMapping({"crear"})
+    public String crear() {
         return "trabajador/nuevo";
     }
 
     @GetMapping(value = "editar")
     public String editar(Model model, @RequestParam("id") String id) {
+        Optional<Trabajador> optionalTrabajador = trabajadorRepository.findById(id);
 
-
-        return "trabajador/editar";
+        if (optionalTrabajador.isPresent()) {
+            Trabajador trabajador= optionalTrabajador.get();
+            model.addAttribute("trabajador", trabajador);
+            return "trabajador/editar";
+        } else {
+            return "redirect:/trabajador/lista";
+        }
     }
-    @PostMapping(value = "/trabajador/guardar")
+    @PostMapping(value = "guardar")
     public String guardar(Trabajador trabajador, RedirectAttributes redirectAttributes) {
         this.trabajadorRepository.save(trabajador);
         redirectAttributes.addFlashAttribute("msg", "Se ha creado un nuevo Trabajador");
         return "redirect:/trabajador/lista";
     }
 
-    @GetMapping(value = "trabajador/borrar")
+    @GetMapping(value = "eliminar")
     public String borrar(@RequestParam("id") String id) {
-        try{
+        Optional<Trabajador> optionalTrabajador = trabajadorRepository.findById(id);
+
+        if (optionalTrabajador.isPresent()) {
             trabajadorRepository.deleteById(id);
-            System.out.println("Se elimino el Trabajador de manera exitosa!");
-        }catch (Exception e) {
-            System.out.println("Se ingreso un id invalido");
         }
         return "redirect:/trabajador/lista";
     }
